nekidaem-test/requirements.txt
Django==3.1.3
psycopg2-binary==2.8.5
djangorestframework~=3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
factory-boy==3.1.0
gunicorn==20.0.4
PyHamcrest==2.0.2
django-tinymce==3.1.0
